+++ src/Randomizer/index.js
@@ -31,6 +31,7 @@ import LudicrousSpeedPatch from './LudicrousSpeedPatch.js';
 import CaseyNoInstantPatch from './CaseyNoInstantPatch.js';
 import CreditsPatch from './CreditsPatch.js';
 import DevmodePatch from './DevmodePatch.js';
+import DropMostRecentPatch from './DropMostRecentPatch.js';
 import ExpandSavePatch from './ExpandSavePatch.js';
 import RunButtonPatch from './RunButtonPatch.js';
 import ShortenPrayerPatch from './ShortenPrayerPatch.js';
@@ -87,6 +88,7 @@ export function execute(romfile, specs, hooks) {
     if(specs.flags.a) patches.push(TrackDoorsPatch);
     if(specs.flags.a) patches.push(TitleDisableGlowPatch);
     if(specs.flags.g >= 1 || specs.flags.s >= 1) patches.push(CaseyNoInstantPatch);
+    if(specs.flags.u >= 1) patches.push(DropMostRecentPatch);
     if(specs.flags.u >= 1) patches.push(RunButtonPatch);
     if(specs.flags.u >= 2) patches.push(LudicrousSpeedPatch);
     if(specs.flags.d >= 3) patches.push(ShortenPrayerPatch);